requirements.txt
langgraph>=0.0.15
langchain>=0.1.0
langchain-anthropic>=0.0.1
typing-extensions>=4.5.0
python-dotenv>=1.0.0
langchain-community
<<<<<<< HEAD
langchain-openai
pypdf
pandas>=2.0.0
langchain-iris
testcontainers-iris
llama-iris
llama-index-legacy
sentence-transformers
langchain
fastembed 
openai 
tiktoken
python-dotenv
pandas
ipykernel
setuptools
=======
pypdf2
pandas>=2.0.0
sentence-transformers>=2.2.2
requests>=2.31.0
tqdm>=4.66.1
numpy>=1.24.0
scikit-learn>=1.3.0
>>>>>>> 6df5799f2bf71db66c86f53bc29e352626b9c4c2
